third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-await-promise-expected.txt

Tests that Runtime.awaitPromise works.

Running test: testResolvedPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            objectId : <number>
            type : object
            value : {
                a : 1
            }
        }
        exceptionId : 0
        lineNumber : 0
        text : Uncaught (in promise)
    }
    result : {
        type : object
        value : {
            a : 1
        }
    }
}

Running test: testRejectedPromiseWithStack
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : <number>
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        text : Uncaught (in promise)
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testPendingPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testResolvedWithoutArgsPromise
{
    result : {
        type : undefined
    }
}

Running test: testGarbageCollectedPromise
{
    code : -32000
    message : Promise was collected
}
